Buttons, Textboxes, and Demo-ready visuals! Whoo!
git-svn-id: https://svn.cs.stolaf.edu/project/sds13-gobbldygook@28 cdf3a981-b4d7-4157-af83-b01e64c43c1b

+++ Color.cpp
@@ -3,10 +3,10 @@
 Color::Color() {
 	red = green = blue = 0;
 }
-Color::Color(int single) {
+Color::Color(double single) {
 	red = green = blue = single;
 }
-Color::Color(int r, int g, int b) {
+Color::Color(double r, double g, double b) {
 	red = r;
 	green = g;
 	blue = b;
