+++ Color.h
@@ -5,10 +5,10 @@
 using namespace std;
 
 struct Color {
-	int red, green, blue;
+	double red, green, blue;
 	Color();
-	Color(int single);
-	Color(int r, int g, int b);
+	Color(double single);
+	Color(double r, double g, double b);
 	Color(const Color &c);
 	void display();
 	Color &operator= (const Color &c);
